Optionalize `FormikErrors`
In the examples given in JS in the docs, A function that is expected to return a `FormikErrors` object creates one with `const errors = {}`, and then conditionally adds only some members to it. This indicates that these members should be optional, which is critical for correct checking in consumer code under `strictNullChecks` mode in typescript.