add test for calling handleSubmit w/o object or preventDefault
In some use cases handleSubmit may be called without providing
an object or the object provided may not have a preventDefault
property. This may occur due to use with custom components,
React Native, or when there is some mechanism that causes form
submission outside and a ReactEvent is not available.

Add test to check that if handleSubmit will not throw an error
if trying to call preventDefault where it does not exist.