Replace lodash.package with lodash/package (#1160)
In this diff I replaced topath and clonedeep packages with submodules in
lodash package. This is more desired choice as in lodash v5 there won't
be single packages. There's also esm version which allows to not
transpile cjs to esm in rollup for example.

lodash-es is used for UMD and ESM bundles.