enableReinitialize note suggestion
At work we had performance issue, we have a pretty big form that was having a hard time to load, around 10 seconds, and one of the main reason was the use of this prop, causing ~10 re-renders on mount, now only 1. so I recommand using this prop to debug your form and remove it after. btw, the other thing we did was adding more memoize functions and selectors (store redux). now the form loads in 2-3 seconds, which is great considering how many values it has.