Fixes broken typing in Formik 2.0.1
Formik returns an array. In TypeScript 3.6.2 (and probably lower), this array is interpreted as being an array of union types (`(FieldProps | FieldMeta)[]`) rather than a tuple (`[FieldProps, FieldMeta]`). `as const` forces the compiler to interpret the array as it is in the literal. This could also be `[FieldProps, FieldMeta]` depending on which version of TS you target.