Fix dirty flag when calling resetForm with current values
When you are resetting the form using `resetForm` and specify current
formik values as the next state the dirty hook won't be run since the
current values didn't change. This adds `intialValues.current` as a
dependency so the hook will run again if either of these changes.

This fixes the dirty part of issue #1976.