Change isSubmitting behaviour to mimic v1
In version 2 the behaviour of isSubmitting was changed from having the
user handle its state after submitting to always resetting it to false
after the onSubmit handler had returned (or it's promise).

This adds back to the version 1 behaviour of not touching the
isSubmitting state after the onSubmit handler has been called if we
don't return a promise in the onSubmit handler.

Solves: #1957