Note about lack of file support
It was my assumption (like many others as per the issue I referenced above) that this library would handle file uploads. Based on this assumption I thought I was using the library wrong when the file wouldn't upload correctly. It would've saved me a lot of time and effort if there was a note somewhere in the docs clearly stating that it cannot handle file uploads. So, this is that note.

This seemed the most obvious place to put it. Let me know if y'all would rather it elsewhere.