Fix useField parameter type
useField was taking FieldAttributes<Val> as argument. If Val is of type
string or other primitive type, FieldAttributes<Val> will be invalid
because primitive type can't be intersected with other type.  In other
words, the following type is invalid.

GenericFieldHTMLAttributes & FieldConfig & string & { name: string }

This results in a type error. When the default type for Val is any. It
broadens the above type to 'any' and hence the default behavior of
useField works when you don't explicitly specify the field's type.