(ts): fix use field typing (#2154)
* Fix useField parameter type

useField was taking FieldAttributes<Val> as argument. If Val is of type
string or other primitive type, FieldAttributes<Val> will be invalid
because primitive type can't be intersected with other type.  In other
words, the following type is invalid.

GenericFieldHTMLAttributes & FieldConfig & string & { name: string }

This results in a type error. When the default type for Val is any. It
broadens the above type to 'any' and hence the default behavior of
useField works when you don't explicitly specify the field's type.

* Fix FieldProps type

V in FieldProps denotes the type of the field, but from was incorrectly
typed to FormikProps<V>.  It should be typed to FormikProps<FormValues>
where FormValues is the type of object that holds the values of form.

To maintain backward compatability, it defaults to any.

* Simplify useField implementation

* Update documentation of useField for proper typing