Bugfix for <Field /> Ignoring the "as" Attribute
I encountered a bug today where "as" had no effect on a Field. It looks like line 134 is *supposed* to take the value of the attribute "as" passed in and assign it to an internal property named "is" which is used later in the code. As written, it instead looks for an input named "is" (which doesn't exist in the docs) and tries to assign it to a property named "as" (which is a reserved word in TS).