Experimenting with batched updates. This slows down constant input significantly (holding down a key).
Moving from a useLayoutEffect to the reducer saves a render for each keystroke, but actually slows down perceived performance by preventing renders from committing until the key is released, whereas the effect would commit about every 5 keys. I checked a bit deeper and it seemed like all the performance was spent in field-level validations' setIn. The majority of CPU time on page was spent in setIn, in fact.