Add note about using React.Fragment within FieldArrray
I generally try to avoid using <div>s where they are not necessary to avoid building huge pyramids in the dom.  
In many instances, replacing <div> tags with <React.Fragment> works without issue.  
Sadly, here it causes the render to fail with a somewhat cryptic error message that does not show where the issue lies and is difficult to debug.

While creating an issue for this would be possible, I am unsure if it is "wanted" behavior and therefore I thought it would be best to just propose a change to the docs, warning about this.