Help to Identify which button is clicked when the user has multiple submit button actions like (Save and Continue, Save, Save for Future, Save and Exit ).
The user passes two arguments to handleSubmit the first parameter is FormEvent and the second parameter is an extra argument.
Users can pass only an extra argument as first a parameter.

The extra argument is passed and 3rd argument to the onSubmit. That helps the user to identify which action he needs to proceed based upon.